### 2021-10-26 #475 @joelnitta
canaper: Categorical Analysis of Neo- And Paleo-Endemism in R
Submitting Author: @joelnitta
Other Package Authors: (delete if none) Shawn Laffan (@shawnlaffan)
Repository:  https://github.com/joelnitta/canaper
Version submitted: 0.0.1
Submission type: Stats
Badge grade: silver
Version accepted: TBD

---

```
Package: canaper
Title: Categorical Analysis of Neo- And Paleo-Endemism
Version: 0.0.1
    c(
    person(given = "Joel H.",
           family = "Nitta",
           email = "joelnitta@gmail.com",
    person(given = "Shawn W.",
           family = "Laffan",
           role = c("ctb", "dtc")),
    person(given = "Brent D.",
           family = "Mishler",
           role = c("ctb", "dtc")),
    person(given = "Wataru",
           family = "Iwasaki",
           role = c("ctb"),
           )
Description: Provides functions to conduct categorical analysis of neo- and paleo-endemism (CANAPE).
Roxygen: list(markdown = TRUE, roclets = c ("collate", "namespace", "rd", "roxyglobals::global_roclet", "srr::srr_stats_roclet"))
    ape,
    assertr,
    assertthat,
    future.apply,
    phyloregion,
    progressr,
    vegan
    future,
    tictoc,
    patchwork,
    tidyverse,
    roxyglobals (>= 0.2.1),
    covr,
    picante
    anthonynorth/roxyglobals
```

- [x] A pre-submission inquiry has been approved in [issue #469](469)

    - Target audience is evolutionary biologists. Scientific application is the analysis of phylogenetic endemism.

    - *The first implementation within* **R** *of an algorithm which has previously been implemented in other languages or contexts*

    - Please include hyperlinked references to all other relevant software: https://github.com/shawnlaffan/biodiverse

    - (Not applicable)

Note also there is a package website available: https://joelnitta.github.io/canaper/

    - Silver, but open to suggestions from the reviewers and editors.

    - Demonstrates excellence in compliance with multiple standards from at least two broad sub-categories
    - Compliance with a good number of standards beyond those identified as minimally necessary. I am in compliance with most of the standards for "General" and "Dimensionality Reduction, Clustering, and Unsupervised Learning".

- [x] I/we have read the [guide for authors](https://devguide.ropensci.org/guide-for-authors.html) and [rOpenSci packaging guide](https://devguide.ropensci.org/building.html).

     - `autotest` hangs (doesn't produce any output after waiting ca. 20 minutes) on most functions. I was only able to use it on `cpr_classify_endem()`.

    - Also, I intend to submit a paper about this package to *Methods in Ecology and Evolution* after ROpenSci review.
