### 2021-10-27 70915e3
Use pre-commit hook instead of codemeta workflow
workflow is consistently failing
