### 2021-10-27 afbd28a
Allow private imports for deps-in-desc hook
