+++ .pre-commit-config.yaml
@@ -64,6 +64,7 @@ repos:
     -   id: parsable-R
     -   id: no-browser-statement
     -   id: deps-in-desc
+        args: ['--allow_private_imports']
 -   repo: https://github.com/pre-commit/pre-commit-hooks
     rev: v4.0.1
     hooks:
