### 2021-10-27 acb4017
Copy functions from phyloregion
to avoid loading betapart.
See https://github.com/HenrikBengtsson/future/discussions/554
Closes #2
