+++ R/utils.R
@@ -247,3 +247,61 @@ match_phylo_comm <- function(phy, comm) {
   res$comm <- comm[, res$phy$tip.label]
   return(res)
 }
+
+# Functions copied from phyloregion v1.0.6 under AGPL-3 ----
+
+# Corresponds to phyloregion::dense2sparse()
+dense2sparse <- function(x) {
+  x <- as.matrix(x)
+  Matrix::Matrix(x, sparse = TRUE)
+}
+
+# Corresponds to phyloregion:::phylo_community()
+phylo_community <- function(x, phy) {
+  el <- numeric(max(phy$edge))
+  el[phy$edge[, 2]] <- phy$edge.length
+  x <- x[, phy$tip.label]
+  anc <- phangorn::Ancestors(phy, seq_along(phy$tip.label))
+  anc <- mapply(c, seq_along(phy$tip.label), anc, SIMPLIFY = FALSE)
+  M <- Matrix::sparseMatrix(as.integer(rep(
+    seq_along(anc),
+    lengths(anc)
+  )), as.integer(unlist(anc)), x = 1L)
+  commphylo <- x %*% M
+  commphylo@x[commphylo@x > 1e-08] <- 1
+  list(Matrix = commphylo, edge.length = el)
+}
+
+# Corresponds to phyloregion::PD()
+PD <- function(x, phy) {
+  if (!methods::is(x, "sparseMatrix")) {
+    stop("x needs to be a sparse matrix!")
+  }
+  if (length(setdiff(colnames(x), phy$tip.label)) > 0) {
+    stop("There are species labels in community matrix missing in the tree!")
+  }
+  if (length(setdiff(phy$tip.label, colnames(x))) > 0) {
+    phy <- ape::keep.tip(phy, intersect(phy$tip.label, colnames(x)))
+  }
+  x <- x[, intersect(phy$tip.label, colnames(x))]
+  z <- phylo_community(x, phy)
+  (z$Matrix %*% z$edge.length)[, 1]
+}
+
+# Corresponds to phyloregion::phylo_endemism()
+phylo_endemism <- function(x, phy, weighted = TRUE) {
+  if (length(setdiff(colnames(x), phy$tip.label)) > 0) {
+    stop("There are species labels in community matrix missing in the tree!")
+  }
+  if (length(setdiff(phy$tip.label, colnames(x))) > 0) {
+    phy <- ape::keep.tip(phy, intersect(phy$tip.label, colnames(x)))
+  }
+  comm_phylo <- phylo_community(x, phy)
+  weights <- comm_phylo$Matrix %*% Matrix::Diagonal(x = 1 / Matrix::colSums(comm_phylo$Matrix))
+  if (weighted == FALSE) {
+    weights[weights < 1] <- 0
+  }
+  pd <- (weights %*% comm_phylo$edge.length)[, 1]
+  pd <- pd[row.names(x)]
+  return(pd)
+}
