+++ R/get_ses.R
@@ -28,8 +28,8 @@
 #'     1:100,
 #'     ~ calc_biodiv_random(comm, phy, phy_alt, "independentswap", 1000L, metrics = "pe")
 #'   )
-#' comm_sparse <- phyloregion::dense2sparse(comm)
-#' pe_obs <- phyloregion::phylo_endemism(comm_sparse, phy, weighted = TRUE)
+#' comm_sparse <- dense2sparse(comm)
+#' pe_obs <- phylo_endemism(comm_sparse, phy, weighted = TRUE)
 #' get_ses(random_vals, pe_obs, "pe")
 #' }
 #' @autoglobal

[…]
