+++ pkgdown.yml
@@ -5,5 +5,5 @@ articles:
   canape: canape.html
   how-many-rand: how-many-rand.html
   parallel: parallel.html
-last_built: 2021-10-26T08:39Z
+last_built: 2021-10-27T03:15Z
 
