### 2021-11-02 #475 @mpadge
@tdhock Only just realised the command there has the wrong syntax - should be `add <@name> to reviewers`, like [this example](https://github.com/ropensci/software-review/issues/470#issuecomment-954890209). Could you please try again? Nevertheless good that you got this one wrong, because it made us realise we need the bot to tell you it didn't understand. Thanks!
