+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: canaper
 Title: Categorical Analysis of Neo- And Paleo-Endemism
-Version: 0.0.1
+Version: 0.0.2
 Authors@R: 
     c(
     person(given = "Joel H.",
