+++ pkgdown.yml
@@ -1,9 +1,9 @@
 pandoc: 2.7.3
-pkgdown: 1.6.1
+pkgdown: 2.0.2
 pkgdown_sha: ~
 articles:
   canape: canape.html
   how-many-rand: how-many-rand.html
   parallel: parallel.html
-last_built: 2021-10-27T03:15Z
+last_built: 2022-01-20T22:28Z
 
