### 2022-02-21 #475 @joelnitta
@KlausVigo thank you very much for the extensive and helpful review!

I just wanted to address one thing first: the reason I copied functions from phyloregion instead of importing them is because of a dependency of phyloregion (betapart), which has a package in `Suggests` (snow) that was resetting the random number generator and thus `cpr_rand_test()` was failing to return identical results when using parallel computing, even when `set.seed()` was used. There is an [issue](https://github.com/joelnitta/canaper/issues/2) describing the problem in canaper, and a related [issue](https://github.com/HenrikBengtsson/future/discussions/554) in the future package.

In the end, I opted to copy the phyloregion functions under the AGPL-3 license, since that fixed the issue and resulted in fewer dependencies. As @HenrikBengtsson [mentioned](https://github.com/HenrikBengtsson/future/discussions/554) though, it seems another way to fix this (and import phyloregion normally) would be to use `future.packages = "snow"` with `future.apply::future_lapply()`. Would you prefer that?
