### 2022-03-08 7e8b249
Revert 'Copy functions from phyloregion'
This reverts commit acb40172c2d5ffb04e13519c6cad7c6d00fa451a.
