+++ tests/testthat/test-cpr_rand_test.R
@@ -19,16 +19,28 @@ dat$comm_mat <- as.matrix(biod_example$comm)
 # For testing that results are same regardless of input type
 # need to have same random seed each time
 set.seed(12345)
-dat$res_from_df <- cpr_rand_test(biod_example$comm, biod_example$phy, null_model = "swap", n_reps = 10)
+dat$res_from_df <- cpr_rand_test(
+  biod_example$comm, biod_example$phy,
+  null_model = "swap", n_reps = 10)
 set.seed(12345)
-dat$res_from_tbl <- cpr_rand_test(dat$comm_tbl, biod_example$phy, null_model = "swap", n_reps = 10, tbl_out = FALSE)
+dat$res_from_tbl <- cpr_rand_test(
+  dat$comm_tbl, biod_example$phy,
+  null_model = "swap", n_reps = 10, tbl_out = FALSE)
 set.seed(12345)
-dat$res_from_mat <- cpr_rand_test(dat$comm_mat, biod_example$phy, null_model = "swap", n_reps = 10)
+dat$res_from_mat <- cpr_rand_test(
+  dat$comm_mat, biod_example$phy,
+  null_model = "swap", n_reps = 10)
 
 # For testing output format
-dat$res_tbl_from_df <- cpr_rand_test(biod_example$comm, biod_example$phy, null_model = "swap", n_reps = 1, metrics = "pd", tbl_out = TRUE)
-dat$res_tbl_from_tbl <- cpr_rand_test(dat$comm_tbl, biod_example$phy, null_model = "swap", n_reps = 1, metrics = "pd")
-dat$res_df_from_tbl <- cpr_rand_test(dat$comm_tbl, biod_example$phy, null_model = "swap", n_reps = 1, metrics = "pd", tbl_out = FALSE)
+dat$res_tbl_from_df <- cpr_rand_test(
+  biod_example$comm, biod_example$phy,
+  null_model = "swap", n_reps = 1, metrics = "pd", tbl_out = TRUE)
+dat$res_tbl_from_tbl <- cpr_rand_test(
+  dat$comm_tbl, biod_example$phy,
+  null_model = "swap", n_reps = 1, metrics = "pd")
+dat$res_df_from_tbl <- cpr_rand_test(
+  dat$comm_tbl, biod_example$phy,
+  null_model = "swap", n_reps = 1, metrics = "pd", tbl_out = FALSE)
 
 # Run tests ----
 
@@ -36,15 +48,19 @@ dat$res_df_from_tbl <- cpr_rand_test(dat$comm_tbl, biod_example$phy, null_model 
 #' @srrstats {G5.0} Use standard dataset for testing (Biodiverse dataset)
 test_that("Input is valid", {
   expect_error(
-    cpr_rand_test(biod_example$comm, biod_example$phy, n_reps = -10, null_model = "swap"),
+    cpr_rand_test(
+      biod_example$comm, biod_example$phy, n_reps = -10, null_model = "swap"),
     "'n_reps' must be > 0"
   )
   expect_error(
-    cpr_rand_test(biod_example$comm, biod_example$phy, n_iterations = -10, null_model = "swap"),
+    cpr_rand_test(
+      biod_example$comm, biod_example$phy,
+      n_iterations = -10, null_model = "swap"),
     "'n_iterations' must be > 0"
   )
   expect_error(
-    cpr_rand_test(10, biod_example$phy, metrics = "pd", null_model = "swap"),
+    cpr_rand_test(
+      10, biod_example$phy, metrics = "pd", null_model = "swap"),
     "'comm' must be of class 'data\\.frame' or 'matrix'"
   )
   expect_error(
@@ -52,7 +68,8 @@ test_that("Input is valid", {
[…]
