+++ R/cpr_rand_test.R
@@ -11,10 +11,11 @@
 #' - `pe`: Phylogenetic endemism (Rosauer et al 2009)
 #' - `rpe`: Relative phylogenetic endemism (Mishler et al 2014)
 #'
-#' (`pe` and `rpe` are needed for CANAPE with \code{\link{cpr_classify_endem}()})
+#' (`pe` and `rpe` are needed for CANAPE with
+#' \code{\link{cpr_classify_endem}()})
 #'
-#' The choice of a randomization algorithm (`null_model`) is not trivial, and may
-#' strongly affect results. `cpr_rand_test()` uses null models provided by
+#' The choice of a randomization algorithm (`null_model`) is not trivial, and
+#' may strongly affect results. `cpr_rand_test()` uses null models provided by
 #' `vegan`; for a complete list, see the help file of [vegan::commsim()] or run
 #' [vegan::make.commsim()]. One frequently used null model is `swap` (Gotelli &
 #' Entsminger 2003), which randomizes the community matrix while preserving
@@ -44,19 +45,23 @@
 #' sites is not recommended generally.
 #'
 #' The following rules apply to `comm` input:
-#' - If dataframe or matrix, must include row names (site names) and column names (species names).
-#' - If tibble, a single column (default, `site`) must be included with site names, and other columns must correspond to species names.
+#' - If dataframe or matrix, must include row names (site names) and column
+#'   names (species names).
+#' - If tibble, a single column (default, `site`) must be included with site
+#'   names, and other columns must correspond to species names.
 #' - Column names cannot start with a number and must be unique.
 #' - Row names (site names) must be unique.
-#' - Values (other than site names) should only include integers >= 0; non-integer input will be converted to integer.
+#' - Values (other than site names) should only include integers >= 0;
+#'   non-integer input will be converted to integer.
 #'
 #' The results are identical regardless of whether the input for `comm` is
 #' abundance or presence-absence data (i.e., abundance weighting is not used).
 #'
-#' @srrstats {G2.0a, G2.1a, G2.3b} Documents expectations on lengths, types of vector
-#'   inputs, case-sensitivity
+#' @srrstats {G2.0a, G2.1a, G2.3b} Documents expectations on lengths,
+#' types of vector inputs, case-sensitivity
 #' @srrstats {G2.7, UL1.0} accept dataframe, tibble, or matrix
-#' @srrstats {UL4.3a} If `tbl_out` is `TRUE`, restricts number of rows/columns printed to screen
+#' @srrstats {UL4.3a} If `tbl_out` is `TRUE`, restricts number of rows/columns
+#' printed to screen
 #' @param comm Dataframe, tibble, or matrix; input community data with
 #'   sites (communities) as rows and species as columns. Either presence-absence
 #'   data (values only 0s or 1s) or abundance data (values >= 0) accepted, but
@@ -81,33 +86,37 @@
 #' contains the site names; only used if `comm` is a tibble (object of class
 #' `tbl_df`).
 #' @param tbl_out Logical vector of length 1; should the output be returned as
-#' a tibble? If `FALSE`, will return a dataframe. Defaults to `TRUE` if `comm` is
-#' a tibble.
+#' a tibble? If `FALSE`, will return a dataframe. Defaults to `TRUE` if `comm`
+#' is a tibble.
 #' @param quiet Logical vector of length 1; if `TRUE`, suppress all warnings and
 #' messages that would be emitted by this function.
 #'
[…]

[…]
