### 2022-03-08 645de44
Remove un-needed conversion to sparse matrix
