+++ tests/testthat/test-cpr_rand_test.R
@@ -459,14 +459,6 @@ test_that("Seeds work across sequential and parallel", {
 
   # Set future resolution to parallelized, with 3 workers
   future::plan(future::multisession, workers = 3)
-  # FIXME: for some bizarre reason, need to "prime" seed with a single run first
-  # need to come up with a better solution
-  par_res_0 <- cpr_rand_test(
-    phylocom$comm, phylocom$phy,
-    null_model = "curveball", n_iterations = 10,
-    n_reps = 10, quiet = TRUE
-  )
-  # then the rest of the set.seed() calls work
   set.seed(12345)
   par_res_1 <- cpr_rand_test(
     biod_example$comm, biod_example$phy,
