### 2022-03-08 #475 @joelnitta
@KlausVigo thanks for making that change: [I switched back to importing phyloregion](https://github.com/joelnitta/canaper/commit/7e8b24950912bfa2a5e10cf4dd9cfa8910d74e76), and it [passes the tests](https://github.com/joelnitta/canaper/commit/0c95b885e76abf563b70099a814d62744f6ee65c) for consistent results using the same seed in parallel.

I just have one request: would you consider bumping the phyloregion version so that I can require the version with this fix in `Imports`? That will prevent this bug popping up for others.

> Hi @joelnitta,
>
> I just made a commit to phyloregion [cb819fb](https://github.com/darunabas/phyloregion/commit/cb819fb430e9c539d36f74937e2c680575d5e0e7). This might solves the problem, at least `sessionInfo()` suggests that `snow` is not loaded via a namespace any more.
>
