+++ pkgdown.yml
@@ -5,5 +5,5 @@ articles:
   canape: canape.html
   how-many-rand: how-many-rand.html
   parallel: parallel.html
-last_built: 2022-01-20T22:28Z
+last_built: 2022-03-08T06:22Z
 
