### 2022-03-09 #475 @KlausVigo
@joelnitta nice that this works and it seems to simplify things.
I wasn't aware that `ImportFrom(betapart, beta.pair)` and `betaprt::beta.pair` made such a large difference with the NAMESPACE. I naively assumed it is about the same and the documentation about these issues does not tell you otherwise. @darunabas is the maintainer of `phyloregion` and has to submit it to CRAN. We probably need a week or so to clean up all changes and get the error messages sorted out as we haven't updated the package for a while.

> @KlausVigo thanks for making that change: [I switched back to importing phyloregion](https://github.com/joelnitta/canaper/commit/7e8b24950912bfa2a5e10cf4dd9cfa8910d74e76), and it [passes the tests](https://github.com/joelnitta/canaper/commit/0c95b885e76abf563b70099a814d62744f6ee65c) for consistent results using the same seed in parallel.
>
> I just have one request: would you consider bumping the phyloregion version so that I can require the version with this fix in `Imports`? That will prevent this bug popping up for others.
