### 2022-03-09 #475 @joelnitta
@KlausVigo no problem: it is a very subtle error, and one that I would never have caught without the help of @HenrikBengtsson. As [he has mentioned](https://twitter.com/henrikbengtsson/status/1460452409776357382) this is likely a much wider problem across many packages on CRAN.

If you and @darunabas are able to release a new version of `phyloregion` in a reasonable amount of time on CRAN that would be great.
