### 2022-07-09 #475 @emilyriederer
Hi @joelnitta and @tdhock  ! I'm currently the rotating editor in chief for rOpenSci submissions. I wanted to check in and see if you expect to have time to revisit the review feedback soon? Otherwise, we may want to put this review on hold status
