+++ .pre-commit-config.yaml
@@ -2,7 +2,7 @@
 # R specific hooks: https://github.com/lorenzwalthert/precommit
 repos:
 -   repo: https://github.com/lorenzwalthert/precommit
-    rev: v0.1.3.9133
+    rev: v0.3.2.9001
     hooks:
     -   id: style-files
         args: [--style_pkg=styler, --style_fun=tidyverse_style]
@@ -12,25 +12,13 @@ repos:
         -    ape
         -    assertr
         -    assertthat
-        -    covr
         -    dplyr
-        -    future
         -    future.apply
-        -    knitr
-        -    magrittr
-        -    patchwork
         -    phyloregion
-        -    picante
         -    progressr
         -    purrr
-        -    rmarkdown
-        -    roxyglobals
         -    stats
-        -    stringr
-        -    testthat
         -    tibble
-        -    tictoc
-        -    tidyverse
         -    vegan
         -    anthonynorth/roxyglobals
         -    ropensci-review-tools/srr
@@ -66,7 +54,7 @@ repos:
     -   id: deps-in-desc
         args: ['--allow_private_imports']
 -   repo: https://github.com/pre-commit/pre-commit-hooks
-    rev: v4.0.1
+    rev: v4.3.0
     hooks:
     -   id: check-added-large-files
         args: ['--maxkb=200']
