### 2022-08-17 b04e2e0
Fix style and typo for vignette
