+++ vignettes/canape.Rmd
@@ -53,7 +53,7 @@ The second, `comm`, is a community dataframe with species as columns and rows as
 
 ```{r inspect-data-comm}
 dim(acacia$comm)
-acacia$comm[1:8,1:8]
+acacia$comm[1:8, 1:8]
 ```
 
 ## Randomization test
@@ -73,7 +73,7 @@ The main purpose of `canaper` is to perform these randomization tests.
 This example also demonstrates one of the strengths of `canaper`: the ability to run randomizations in parallel^[For more information on how and when to use parallel computing in `canaper`, [see the "Parallel computing" vignette](https://joelnitta.github.io/canaper/articles/parallel.html)]. 
 This is by far the most time-consuming part of CANAPE, since we have to repeat the calculations many (e.g., hundreds or more) times across the randomized communities to obtain reliable results. Here, we set the number of iterations (`n_iterations`; i.e., the number of swaps used to produce each randomized community) fairly high because this community matrix is large and includes many zeros; thorough mixing by swapping many times is required to completely randomize the matrix. 
 
-We will use a low number of random communities (`n_reps`) so things finish relatively quickly; you should consider increasing `n_reps` for a "real" analysis^[For more information on setting the appropriate number of iterations and replicates, [see the "How many randomizations?" vigenette](https://joelnitta.github.io/canaper/articles/how-many-rand.html).].
+We will use a low number of random communities (`n_reps`) so things finish relatively quickly; you should consider increasing `n_reps` for a "real" analysis^[For more information on setting the appropriate number of iterations and replicates, [see the "How many randomizations?" vignette](https://joelnitta.github.io/canaper/articles/how-many-rand.html).].
 We will use the `curveball` randomization algorithm, which maintains species richness and abundance patterns while randomizing species identity [@Strona2014]^[`curveball` is similar to `swap` but runs much faster so I chose it for this large dataset.].
 
 ```{r acacia-rand}
@@ -81,18 +81,20 @@ We will use the `curveball` randomization algorithm, which maintains species ric
 plan(multisession, workers = 3)
 
 # Uncomment this to show a progress bar when running cpr_rand_test()
-# progressr::handlers(global = TRUE)
+# progressr::handlers(global = TRUE) # nolint
 
-# Set a random number generator seed so we get the same results if this is run again
+# Set a random number generator seed so we get the same results if this is
+# run again
 set.seed(071421)
 
 tic() # Set a timer
 # Run randomization test
 acacia_rand_res <- cpr_rand_test(
-  acacia$comm, acacia$phy, 
+  acacia$comm, acacia$phy,
   null_model = "curveball",
   n_reps = 100, n_iterations = 100000,
-  tbl_out = TRUE)
+  tbl_out = TRUE
+)
 toc() # See how long it took
 
 # Switch back to sequential (non-parallel) mode
@@ -129,7 +131,7 @@ acacia_canape <- cpr_classify_endem(acacia_rand_res)
 table(acacia_canape$endem_type)
 ```
 
-## Classify signficance
+## Classify significance
 
[…]
