### 2022-08-17 d8aaf8d
Update documentation
- Add list of similar packages
- Link to CONTRIBUTING.md, CoC
