+++ DESCRIPTION
@@ -24,7 +24,7 @@ License: MIT + file LICENSE
 Encoding: UTF-8
 LazyData: true
 Roxygen: list(markdown = TRUE, roclets = c ("collate", "namespace", "rd", "roxyglobals::global_roclet", "srr::srr_stats_roclet"))
-RoxygenNote: 7.1.2
+RoxygenNote: 7.2.0
 URL: https://github.com/joelnitta/canaper
 BugReports: https://github.com/joelnitta/canaper/issues
 Imports: 
@@ -54,7 +54,7 @@ Suggests:
     picante
 Config/testthat/edition: 3
 Depends: 
-    R (>= 3.5.0)
+    R (>= 4.1.0)
 VignetteBuilder: knitr
 Remotes: 
     anthonynorth/roxyglobals
