### 2022-08-18 f3af467
Decrease n_reps in vignettes
