+++ vignettes/parallel.Rmd
@@ -43,12 +43,13 @@ First, let's run an analysis in default sequential mode (without parallel comput
 ```{r sequential-1}
 tic()
 biod_res_seq <- cpr_rand_test(
-  biod_example$comm, biod_example$phy, 
-  null_model = "swap", n_reps = 50)
+  biod_example$comm, biod_example$phy,
+  null_model = "swap", n_reps = 50
+)
 toc()
 ```
 
-Since we have specified 100 random replicates and are not using parallel computing, `cpr_rand_test()` calculated the various phylogenetic diversity metrics for each of the 100 replicates one at a time.
+Since we have specified 50 random replicates and are not using parallel computing, `cpr_rand_test()` calculated the various phylogenetic diversity metrics for each of the 50 replicates one at a time.
 
 ### Parallel mode
 
@@ -60,23 +61,24 @@ availableCores()
 
 OK, we have verified that there are multiple cores available for parallel computing.
 
-To enable parallel computing, just add one line before `cpr_rand_test()`: `plan(multisession, workers = 3)` ^[Of course, the number of workers should be no greater than the number of available CPUs.]. Here, the `multisession, workers = 3` part is telling `future` that we want to use 3 CPUs in parallel on our local machine. See `future::plan()` for other options. Otherwise, **everything is the same**.
+To enable parallel computing, just add one line before `cpr_rand_test()`: `plan(multisession, workers = 2)` ^[Of course, the number of workers should be no greater than the number of available CPUs.]. Here, the `multisession, workers = 2` part is telling `future` that we want to use 2 CPUs in parallel on our local machine. See `future::plan()` for other options. Otherwise, **everything is the same**.
 
 ```{r parallel-1}
-# Set up parallel computing with 3 CPUs
-plan(multisession, workers = 3)
+# Set up parallel computing with 2 CPUs
+plan(multisession, workers = 2)
 
 tic()
 biod_res_par <- cpr_rand_test(
-  biod_example$comm, biod_example$phy, 
-  null_model = "swap", n_reps = 50)
+  biod_example$comm, biod_example$phy,
+  null_model = "swap", n_reps = 50
+)
 toc()
 
 # Change back to default sequential mode
 plan(sequential)
 ```
 
-This time, the calculations were carried out in 3 batches in parallel.
+This time, the calculations were carried out in 2 batches in parallel.
 
 But there is no significant improvement in processing time^[Results vary each time this vignette is generated, but it is usually about the same; sometimes, the parallel version is even **slower**.]. What is going on here?
 
@@ -108,22 +110,24 @@ Let's see how parallel computing works on the `acacia` dataset:
 plan(sequential)
 tic()
 acacia_res_seq <- cpr_rand_test(
-  acacia$comm, acacia$phy, 
-  null_model = "swap", n_reps = 50)
+  acacia$comm, acacia$phy,
+  null_model = "curveball", n_reps = 100
+)
 toc()
 ```
 
 ```{r acacia-parallel}
-# Run cpr_rand_test() in parallel with 3 CPUs
-plan(multisession, workers = 3)
+# Run cpr_rand_test() in parallel with 2 CPUs
+plan(multisession, workers = 2)
 tic()
 acacia_par_seq <- cpr_rand_test(
-  acacia$comm, acacia$phy, 
-  null_model = "swap", n_reps = 50)
+  acacia$comm, acacia$phy,
+  null_model = "curveball", n_reps = 100
+)
 toc()
 ```
 
-And now we start to see the performance improvements that be can be gained from parallel computing!
[…]
