+++ vignettes/how-many-rand.Rmd
@@ -71,21 +71,23 @@ dist_test_res <- tibble(
   n_iter = magrittr::raise_to_power(10, 1:6),
   # Make one random community for each value of `n_iter`
   rand_comm = map(
-    n_iter, 
-    ~cpr_rand_comm(
-      comm_bin, 
-      null_model = "swap", n_iterations = .)),
+    n_iter,
+    ~ cpr_rand_comm(
+      comm_bin,
+      null_model = "swap", n_iterations = .
+    )
+  ),
   # Calculate the Pearson correlation (r) between each randomized
   # matrix and the original matrix by converting
   # the matrices to vectors
   corr = map_dbl(
-    rand_comm, 
-    ~cor(c(comm_bin), c(.)) #c() converts a matrix to vector
+    rand_comm,
+    ~ cor(c(comm_bin), c(.)) # c() converts a matrix to vector
   )
 )
 
 # Plot the results
-ggplot(dist_test_res, aes(x = n_iter, y = corr)) + 
+ggplot(dist_test_res, aes(x = n_iter, y = corr)) +
   geom_point() +
   geom_line() +
   scale_x_log10() +
@@ -108,25 +110,31 @@ Here, I will compare the percentile of observed phylogenetic diversity relative 
 # Specify a different random seed for each set of randomizations so they give
 # different, reproducible results
 
+# Set a parallel back-end, with 2 CPUs running simultaneously to speed things up
+plan(multisession, workers = 2)
+
 # First set
 set.seed(12345)
 tic()
 res_10_1 <- cpr_rand_test(
-  biod_example$comm, biod_example$phy, 
+  biod_example$comm, biod_example$phy,
   null_model = "swap",
-  n_iterations = 10000, n_reps = 10, tbl_out = TRUE)
+  n_iterations = 10000, n_reps = 10, tbl_out = TRUE
+)
 toc()
 tic()
 res_100_1 <- cpr_rand_test(
-  biod_example$comm, biod_example$phy, 
+  biod_example$comm, biod_example$phy,
   null_model = "swap",
-  n_iterations = 10000, n_reps = 100, tbl_out = TRUE)
+  n_iterations = 10000, n_reps = 100, tbl_out = TRUE
+)
 toc()
 tic()
 res_1000_1 <- cpr_rand_test(
-  biod_example$comm, biod_example$phy, 
+  biod_example$comm, biod_example$phy,
   null_model = "swap",
-  n_iterations = 10000, n_reps = 1000, tbl_out = TRUE)
+  n_iterations = 10000, n_reps = 1000, tbl_out = TRUE
+)
 toc()
 
 # Second set
@@ -134,15 +142,21 @@ set.seed(67890)
 res_10_2 <- cpr_rand_test(
   biod_example$comm, biod_example$phy,
   null_model = "swap",
-  n_iterations = 10000, n_reps = 10, tbl_out = TRUE)
+  n_iterations = 10000, n_reps = 10, tbl_out = TRUE
+)
 res_100_2 <- cpr_rand_test(
-  biod_example$comm, biod_example$phy, 
+  biod_example$comm, biod_example$phy,
   null_model = "swap",
-  n_iterations = 10000, n_reps = 100, tbl_out = TRUE)
+  n_iterations = 10000, n_reps = 100, tbl_out = TRUE
+)
 res_1000_2 <- cpr_rand_test(
-  biod_example$comm, biod_example$phy, 
+  biod_example$comm, biod_example$phy,
   null_model = "swap",
-  n_iterations = 10000, n_reps = 1000, tbl_out = TRUE)
+  n_iterations = 10000, n_reps = 1000, tbl_out = TRUE
+)
+
[…]
