+++ DESCRIPTION
@@ -24,7 +24,7 @@ License: MIT + file LICENSE
 Encoding: UTF-8
 LazyData: true
 Roxygen: list(markdown = TRUE, roclets = c ("collate", "namespace", "rd", "roxyglobals::global_roclet", "srr::srr_stats_roclet"))
-RoxygenNote: 7.2.0
+RoxygenNote: 7.2.1
 URL: https://github.com/joelnitta/canaper
 BugReports: https://github.com/joelnitta/canaper/issues
 Imports: 
