### 2022-08-18 bccd027
Remove un-needed checks
from cpr_rand_comm() and calc_biodiv_random(). Fixes #5
