+++ tests/testthat/test-calc_biodiv_random.R
@@ -9,39 +9,6 @@ phy$edge.length <- phy$edge.length / sum(phy$edge.length) # nolint
 
 # Run tests ----
 
-#' @srrstats {G5.2, G5.2a, G5.2b, UL7.0} tests failure if input is not valid
-# and checks warning messages
-test_that("Input is valid", {
-  expect_error(
-    calc_biodiv_random(10, phy, phy_alt, "swap", 10L),
-    "'comm' must be of class 'data\\.frame' or 'matrix'"
-  )
-  expect_error(
-    calc_biodiv_random(comm, NA, phy_alt, "swap", 10L),
-    "'phy' must be a list of class 'phylo'"
-  )
-  expect_error(
-    calc_biodiv_random(comm, phy, phy_alt, "swap", 10L, metrics = "pg"),
-    "'metrics' may only include 'pd', 'rpd', 'pe', 'rpe', 'pd_alt', 'pe_alt'"
-  )
-  expect_error(
-    calc_biodiv_random(comm, phy, phy_alt, "swar", 10L, metrics = "pd"),
-    "'null_model' must be one of"
-  )
-  expect_error(
-    calc_biodiv_random(comm, phy, phy, "swap", 10L),
-    "'phy' and 'phy_alt' should not be identical"
-  )
-  expect_error(
-    calc_biodiv_random(comm, biod_example$phy, phy_alt, "swap", 10L),
-    "phylogeny not rescaled to total length 1"
-  )
-  expect_error(
-    calc_biodiv_random(comm, phy, biod_example$phy, "swap", 10L),
-    "alternative phylogeny not rescaled to total length 1"
-  )
-})
-
 test_that("Random seeds work", {
   set.seed(12345)
   res1 <- calc_biodiv_random(comm, phy, phy_alt, "curveball", 100L)
