### 2022-08-18 12bbe36
Decrease number of reps and cores in tests and examples
