+++ man/cpr_classify_endem.Rd
@@ -37,7 +37,7 @@ set.seed(12345)
 data(phylocom)
 rand_test <- cpr_rand_test(
   phylocom$comm, phylocom$phy,
-  null_model = "curveball", metrics = c("pe", "rpe")
+  null_model = "curveball", metrics = c("pe", "rpe"), n_reps = 50
 )
 cpr_classify_endem(rand_test)
 }
