### 2022-08-18 93543d3
Add option to slince match_phylo_comm warnings
Fixes #4
