### 2022-08-18 #475 @joelnitta
## Response to reviews

I am sorry for taking so long to respond to the reviewers. Thanks again to both reviewers for their helpful comments. Below I respond to each comment as appropriate.

### Response to @luismurao

- [x] Concern about parallelization. I think it would be better to have parameters for parallelization in functions that allow running analysis in parallel (so calls are done inside them).
- **Response:** The second reviewer, @KlausVigo disagreed with this point and was in favor of leaving the current approach to parallelization, so I am making no changes.

- [x] Instead of suggesting the `future` package I would import it
- **Response:** I disagree with this. Parallelization is not required, it is an option to speed up analyses if desired. So I think it makes sense as `Suggests` instead of `Imports`. No changes made.

- [x] No code of conduct and contribution guidelines, please add it
- **Response:** There was apparently a misunderstanding; contribution guidelines are already provided at `/.github/CONTRIBUTING.md`. To make it easier to find though, I have added a link to this document from the README. I also added a link to the ROpenSci code of conduct using the [suggested wording in the developer guide](https://devguide.ropensci.org/collaboration.html?q=code%20of%20c#coc-file) ([d8aaf8de5193166d52ab890d8267090ea2f8ef6a](https://github.com/joelnitta/canaper/commit/d8aaf8de5193166d52ab890d8267090ea2f8ef6a)).

- [x] Please change parameter Depends: R (>= 3.5.0) to Depends: R (>= 4.1.0) in the description file or provide a way to install an earlier version of phangorn (version 2.5.5 should work).
- **Response:** I changed to `Depends: R (>= 4.1.0)` ([0f334965fb5a3fe5034eb1bc682b5051548eddb1](https://github.com/joelnitta/canaper/commit/0f334965fb5a3fe5034eb1bc682b5051548eddb1)).

#### Automated test

- [x] `devtools::check()` takes more than six minutes. Building of vignette outputs takes a lot of time (5m 5.2s). Consider using lees number of replicates in examples (i.e. `n_reps` of `canaper::cpr_rand_test`).
- **Response:** I had to include a non-trivial number replicates to produce reasonable results (e.g., in comparison with Mishler et al. 2014 for the [CANAPE example vignette](https://joelnitta.github.io/canaper/articles/canape.html)) and to demonstrate performance gains from parallelization. But I was able to decrease the number of replicates in the vignettes ([f3af4671d679020aeaa97b1a038b3b5a96633f42](https://github.com/joelnitta/canaper/commit/f3af4671d679020aeaa97b1a038b3b5a96633f42)). The vignettes now take 3m 40s on my computer to compile.

- [x] I got one error after upgrading from GEOS 3.9 to 3.10 in Linux. The issue is related to phyloregion dependency.
- **Response:** Yes: phylogregion imports `rgeos`, which is [going to be retired soon](https://r-spatial.org/r/2022/04/12/evolution.html) and [needs fixing](https://github.com/darunabas/phyloregion/issues/11). Correct installation of external dependencies to phyloregion is outside the scope of `canaper`. No change made (for now).

- [x] `devtools::test()` finished with no fails and one warning
- **Response:** On my computer, `devtools::test()` now finishes with no fails or warnings (at [93543d399fd34dd6546e56368986a8afa133f190](https://github.com/joelnitta/canaper/commit/93543d399fd34dd6546e56368986a8afa133f190)).

- [x] `devtools::spell_check()` Change "signficance" to "significance" in line 132
- **Response:** Fixed ([b04e2e0a8397568bcc615e4aaeb5a0d9b02ad85e](https://github.com/joelnitta/canaper/commit/b04e2e0a8397568bcc615e4aaeb5a0d9b02ad85e))

#### Performance

- [x] When benchmarking, I found a strange behavior. After using more than four threads, the program takes more time to finish a randomization test.
- **Response:** Speed is not expected to increase linearly with the number of cores. It is well documented in other software that increasing cores above a certain point causes code to run more slowly due to the resources required to coordinate threads. It is up to the user to identify the optimal number of threads to use. I provide a vignette article discussing precisely this topic: https://joelnitta.github.io/canaper/articles/parallel.html.

#### Questions and comments

- [x] I was wondering if canaper has already all methods of the CANAPE version of Biodiverse. Could write something about it in the README?
- **Response:** I have added a section in the README comparing `canaper` other software ([d8aaf8de5193166d52ab890d8267090ea2f8ef6a](https://github.com/joelnitta/canaper/commit/d8aaf8de5193166d52ab890d8267090ea2f8ef6a)).

- [x] Although the vignettes are very useful and well documented, it would be nice to have methods for plotting spatial data results.
- **Response:** Thanks for the suggestion. I have already included [color palettes](https://joelnitta.github.io/canaper/reference/index.html#color-palettes) for plotting results that should be accessible regardless of color vision type. IMHO there are too many customizations that users would want to make to try and implement a one-size-fits-all plotting method. Rather, the results are structured in a way that they should be easily plotted with e.g. ggplot2, and I [demonstrate this in the vignettes](https://joelnitta.github.io/canaper/articles/canape.html#visualize-results).

### Response to @KlausVigo

- [x] There are some functions in the `picante` (`ses.pd`) and in the `PhyloMeasures` package for Faith's phylogenetic diversity with a similar scope.
- **Response:** There are many R packages that calculate community diversity metrics, and many of them overlap in common metrics like Faith's PD. In the case of `canaper`, I need to calculate Faith's PD as part of the CANAPE calculation. I use the `phyloregion::PD()` to do this as it is the fastest AFAIK. No changes needed.

- [x] The canaper package builds on functions from the vegan package to resample (dense) community matrices and borrows biodiversity measures (Faith's phylogenetic diversity and phylogenetic endemism) from the phyloregion package. The picante package also has some functions for swapping (community) matrices similar to vegan. A comparison between the packages would be useful.
- **Response:** I have added a section to the README comparing `canaper` with other software ([d8aaf8de5193166d52ab890d8267090ea2f8ef6a](https://github.com/joelnitta/canaper/commit/d8aaf8de5193166d52ab890d8267090ea2f8ef6a)).

#### Specific comments

- [x] The functions from the phyloregion package used in the package should be acknowledged properly... An alternative is importing the functions from the package directly.
- **Response:** Please see response [above](https://github.com/ropensci/software-review/issues/475#issuecomment-1047206275). Functions from the phyloregion package are now imported, not copied ([7e8b24950912bfa2a5e10cf4dd9cfa8910d74e76](https://github.com/joelnitta/canaper/commit/7e8b24950912bfa2a5e10cf4dd9cfa8910d74e76)). Please note that phyloregion is still at [v1.0.7 as of writing](https://github.com/darunabas/phyloregion/tree/9b47a325054e9a1b801598a0469e472eefa6be53) (same version as before @KlausVigo [implemented the change](https://github.com/ropensci/software-review/issues/475#issuecomment-1060658515) that fixed the loading of `snow`), so I can't specify the version of phyloregion that would avoid canaper [issue #2](https://github.com/joelnitta/canaper/issues/2).

- [x] In contrast to the other reviewer I really like the use of the `future` package.
- **Response:** I agree, and have left parallelization as it is.

#### Profiling

- [x] Providing a swapping algorithm which can handle sparse matrices would be a huge improvement for the community.
- **Response:** I agree implementing sparse matrix encoding for generating random matrices would be a huge benefit and will consider it as a feature to be added possibly in the future (any help would be appreciated!).

- [x] The functions `calc_biodiv_random` and `cpr_rand_comm` are only called internally and are called thousands of times inside of the loop. It should be save to remove all (most) of the calls to assertthat::assert_that from these functions as parameters have been checked already and are save to use!
- **Response:** Actually, `cpr_rand_comm()` is exported, but point well-taken. I have deleted the redundant checks ([bccd0278dd3de5853680d9979ca26c1aa38a7fb3](https://github.com/joelnitta/canaper/commit/bccd0278dd3de5853680d9979ca26c1aa38a7fb3)).

#### For submitting to CRAN

- [x] There are some instances where more than 2 cores are used. This could fail on CRAN, when submitting.
- **Response:** Thanks for pointing this out. Fixed ([12bbe36bf5516a2a63c1235c76234708c4c5432d](https://github.com/joelnitta/canaper/commit/12bbe36bf5516a2a63c1235c76234708c4c5432d), [f3af4671d679020aeaa97b1a038b3b5a96633f42](https://github.com/joelnitta/canaper/commit/f3af4671d679020aeaa97b1a038b3b5a96633f42))

- Both reviewers left `Performance: Any performance claims of the software been confirmed` unchecked. I am not sure why this is the case, as I include tests of increase in performance using parallel computing **and** demonstrate this feature in a [vignette](https://joelnitta.github.io/canaper/articles/parallel.html).

- I have added @KlausVigo as a reviewer, but not @luismurao, according to their responses to `Should the author(s) deem it appropriate, I agree to be acknowledged as a package reviewer`.
