### 2022-08-26 3037d55
Run usethis::use_pkgdown_github_pages() after updating usethis to 2.1.6
