### 2022-08-26 f8e2dbe
Load future package in vignette
