+++ .github/workflows/pkgdown.yaml
@@ -30,7 +30,7 @@ jobs:
 
       - uses: r-lib/actions/setup-r-dependencies@v2
         with:
-          cache-version: 2
+          cache-version: 1 # change to any other number than 1 to clear cache
           extra-packages: any::pkgdown, local::.
           needs: website
 
