### 2022-08-26 8dd9752
Update package Rd file
