+++ man/canaper-package.Rd
@@ -12,6 +12,7 @@ Provides functions to conduct categorical analysis of neo- and paleo-endemism (C
 Useful links:
 \itemize{
   \item \url{https://github.com/joelnitta/canaper}
+  \item \url{https://joelnitta.github.io/canaper/}
   \item Report bugs at \url{https://github.com/joelnitta/canaper/issues}
 }
 
