+++ tests/testthat/test-utils.R
@@ -159,3 +159,13 @@ test_that("Matching community and phylogeny data works", {
     fixed = TRUE
   )
 })
+
+test_that("skip_extended_if_no_commit works", {
+  # Test that a skip happens
+  withr::local_envvar(CANAPER_EXTENDED_TESTS = "true")
+  expect_condition(skip_extended_if_no_commit(), class = "skip")
+
+  # Test that a skip doesn't happen
+  withr::local_envvar(CANAPER_EXTENDED_TESTS = "")
+  expect_condition(skip_extended_if_no_commit(), NA, class = "skip")
+})
