### 2022-08-26 2bb651f
Increase number of reps for testing Parallelization decreases calculation time
