+++ pkgdown.yml
@@ -5,7 +5,7 @@ articles:
   canape: canape.html
   how-many-rand: how-many-rand.html
   parallel: parallel.html
-last_built: 2022-08-26T09:40Z
+last_built: 2022-08-26T22:44Z
 urls:
   reference: https://joelnitta.github.io/canaper/reference
   article: https://joelnitta.github.io/canaper/articles
