### 2022-08-27 7df09cd
Change na.rm to na_rm in count_*()
