### 2022-08-27 6818499
Fix extended test settings
