+++ .github/workflows/R-CMD-check.yaml
@@ -31,7 +31,7 @@ jobs:
       R_REMOTES_NO_ERRORS_FROM_WARNINGS: true
       RSPM: ${{ matrix.config.rspm }}
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
-      MYPKG_EXTENDED_TESTS: ${{contains(github.event.head_commit.message, 'run-extended')}}
+      CANAPER_EXTENDED_TESTS: ${{contains(github.event.head_commit.message, 'run-extended')}}
 
     steps:
       - uses: actions/checkout@v2
