+++ pkgdown.yml
@@ -5,7 +5,7 @@ articles:
   canape: canape.html
   how-many-rand: how-many-rand.html
   parallel: parallel.html
-last_built: 2022-08-27T01:40Z
+last_built: 2022-08-27T02:05Z
 urls:
   reference: https://joelnitta.github.io/canaper/reference
   article: https://joelnitta.github.io/canaper/articles
