### 2022-08-27 #475 @joelnitta
## Update to response to reviewers

The longest running test is [`"Parallelization decreases calculation time"`](https://github.com/joelnitta/canaper/blob/76c277490dcecd0f32df351c48c8f8c891674aad/tests/testthat/test-cpr_rand_test.R#L391). To make this test robust, I have to include a relatively high number of replicates, otherwise by random chance sometimes running in parallel won't be faster than running sequentially.

I have decided to follow [6.1.5.4 Extended tests](https://stats-devguide.ropensci.org/standards.html#extended-tests) of [rOpenSci Statistical Software Peer Review](https://stats-devguide.ropensci.org/standards.html) and treat this as an extended test that can be optionally skipped (it is now skipped by default).

I have [documented this in a comment](https://github.com/joelnitta/canaper/blob/76c277490dcecd0f32df351c48c8f8c891674aad/tests/testthat/test-cpr_rand_test.R#L395) as follows:

```
  # This test is skipped by default because of its long run time.
  # To include the test locally, first run
  # withr::local_envvar(CANAPER_EXTENDED_TESTS = "true") # nolint
  # To include the test in CI (github actions), include the phrase
  # 'run-extended' in the commit
```

`devtools::test()` on my computer now finishes in 36 s skipping the extended test, and 181 s when including it.

@tdhock, @emilyriederer sorry again for the long delay, but this should be ready to move forward. Please let me know if there is anything you need from my end.
