### 2022-08-29 #475 @emilyriederer
Also tagging  @KlausVigo @luismurao . I know this review was a bit delayed and your availability may have changed. At your convenience, if you could please review @joelnitta 's responses and share whether these address your feedback ([template available here](https://devguide.ropensci.org/reviewerguide.html#followupreviewer)) , that would be fantastic.
