+++ data-raw/acacia.R
@@ -4,13 +4,17 @@ library(janitor)
 library(ape)
 library(tidyverse)
 library(assertr)
+library(assertthat)
 library(here)
 
 # Load community ----
 
-# acacia_sites_by_spp.csv downloaded from github
-# https://github.com/shawnlaffan/biodiverse/tree/73522b74e52a5fb77ae5cfc2e90010350a3abf70/etc/experiments/independent_swaps
-acacia_comm <- read_csv(here("data-raw/acacia_sites_by_spp.csv"), col_types = cols(.default = col_character())) %>%
+# Community data: Acacia community data from Mishler et al. 2014
+# For more info, see data-raw/CANAPE_Acacia/README.md
+acacia_comm <- read_csv(
+  here("data-raw/CANAPE_Acacia/sites_by_spp.csv"),
+  col_types = cols(.default = col_character())
+) %>%
   clean_names() %>%
   select(-axis_0, -axis_1) %>%
   mutate(across(-element, as.numeric)) %>%
@@ -19,11 +23,20 @@ acacia_comm <- read_csv(here("data-raw/acacia_sites_by_spp.csv"), col_types = co
 
 # Load tree ----
 
-# Need to do extra parsing of nexus file, since taxa names include spaces, parentheses, etc.
+# Need to do extra parsing of nexus file, since taxa names include spaces,
+# parentheses, etc.
 temp_phy <- tempfile("temp_phy.tre")
 
-acacia_nexus_raw <- readr::read_lines(here("data-raw/1_1363828941_Acacia.nexorg"))
+# Read in tree nexus file from treebase
+acacia_nexus_raw <- readr::read_lines(
+  here("data-raw/1_1363828941_Acacia.nexorg")
+)
 
+# The nexus file has one block for taxa names and
+# one block for the tree.
+# The tips of the tree are labeled with number codes, which
+# correspond to taxon names.
+# We want to relabel the tree so the tips are the taxon names, not numbers
 acacia_nexus_raw %>%
   magrittr::extract(str_detect(., "TREE tree_1 = ")) %>%
   unlist() %>%
@@ -44,26 +57,51 @@ taxa_end <- acacia_nexus_raw %>%
   which() %>%
   magrittr::subtract(1)
 
-acacia_taxa <- acacia_nexus_raw[taxa_start:taxa_end] %>%
-  readr::read_table(col_names = FALSE) %>%
-  separate(X1, c("number", "name"), sep = " ", extra = "merge") %>%
+# Write taxa block to temporary file to read in with read_tsv
+temptable <- tempfile("table.tsv")
+
+writeLines(acacia_nexus_raw[taxa_start:taxa_end], temptable)
+
+# Make table of taxon number and name for all Acacia plus outgroups
+acacia_with_og <- readr::read_tsv(
+  temptable,
+  col_names = FALSE, quote = "",
+  col_types = cols(.default = col_character())
+) %>%
+  separate(X3, c("number", "name"), sep = " ", extra = "merge") %>%
+  select(number, name)
+
+unlink(temptable)
+
+# Format table of ingroup taxa (named with specific epithet only)
+acacia_taxa <-
+  acacia_with_og %>%
+  # drop outgroups
   filter(str_detect(name, "Acacia")) %>%
   mutate(name = str_replace_all(name, " ", "_")) %>%
   extract(name, "species", "(Acacia_[^_]+)", remove = FALSE) %>%
-  mutate(species = str_remove_all(species, "Acacia_") %>% str_remove_all(",")) %>%
+  mutate(
+    species = str_remove_all(species, "Acacia_") %>%
+      str_remove_all(",")
+  ) %>%
[…]
