### 2022-09-14 #475 @ropensci-review-bot
Approved! Thanks @joelnitta for submitting and @KlausVigo, @luismurao for your reviews! :grin:
